Create service for check url, take all http redirect code as success
